+++ src/main/java/woowacourse/globalException/ControllerAdvice.java
@@ -1,7 +1,9 @@
 package woowacourse.globalException;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -9,9 +11,15 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import woowacourse.auth.exception.InvalidTokenException;
 import woowacourse.globalException.dto.ErrorResponse;
 import woowacourse.member.exception.MemberException;
+import woowacourse.shoppingcart.exception.InvalidCartItemException;
+import woowacourse.shoppingcart.exception.InvalidOrderException;
+import woowacourse.shoppingcart.exception.InvalidProductException;
+import woowacourse.shoppingcart.exception.NotInMemberCartItemException;
+
+import javax.validation.ConstraintViolationException;
 
 @RestControllerAdvice
-public class GlobalControllerAdvice {
+public class ControllerAdvice {
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity<ErrorResponse> parameterException(MethodArgumentNotValidException e) {
@@ -31,6 +39,34 @@ public class GlobalControllerAdvice {
     @ExceptionHandler(InvalidTokenException.class)
     public ResponseEntity<ErrorResponse> handleInvalidMemberException(InvalidTokenException e) {
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse(e.getMessage()));
+    }
+
+    @ExceptionHandler(EmptyResultDataAccessException.class)
+    public ResponseEntity handle() {
+        return ResponseEntity.badRequest().body("존재하지 않는 데이터 요청입니다.");
+    }
+
+    @ExceptionHandler({
+            HttpMessageNotReadableException.class,
+            ConstraintViolationException.class,
+    })
+    public ResponseEntity handleInvalidRequest(final RuntimeException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler({
+            InvalidCartItemException.class,
+            InvalidProductException.class,
+            InvalidOrderException.class,
+            NotInMemberCartItemException.class,
+    })
+    public ResponseEntity handleInvalidAccess(final RuntimeException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler(RuntimeException.class)
+    public ResponseEntity handleUnhandledException() {
+        return ResponseEntity.badRequest().body("Unhandled Exception");
     }
 
     @ExceptionHandler(Exception.class)
